# menu/openapi.snapshot.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "TAWROS API",
        "version": "1.0.0",
        "description": "API REST para frontend TAWROS, integrações e automações"
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
...
    },
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "summary": "Login com e-mail e senha",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "summary": "Renova access token usando refresh token",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "refreshToken"
                                ],
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "minLength": 20
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "summary": "Revoga refresh token",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "refreshToken"
                                ],
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "minLength": 20
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/auth/revoke": {
            "post": {
                "summary": "Revoga refresh token (admin/user autenticado)",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "refreshToken"
                                ],
                                "properties": {
                                    "refreshToken": {
                                        "type": "string",
                                        "minLength": 20
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "summary": "Dados do usuario autenticado",
                "tags": [
                    "Auth"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/auth/change-password": {
            "post": {
                "summary": "Troca senha com validação da senha atual",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "currentPassword",
                                    "newPassword"
                                ],
                                "properties": {
                                    "currentPassword": {
                                        "type": "string",
                                        "minLength": 6
                                    },
                                    "newPassword": {
                                        "type": "string",
                                        "minLength": 8
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/auth/forgot-password": {
            "post": {
                "summary": "Solicita reset por e-mail, CPF ou celular",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "cpf": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
        "/api/v1/auth/reset-password": {
            "post": {
                "summary": "Reset de senha com token",
                "tags": [
                    "Auth"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "token",
                                    "newPassword"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string"
                                    },
                                    "newPassword": {
                                        "type": "string",
                                        "minLength": 8
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Default Response"
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://192.168.15.26:3000"
        }
    ]
}